Refactored login.js  and signup.js files to controllers
